requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
assemblyai==0.36.0
attrs==25.1.0
autopep8==2.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
distro==1.9.0
Django==4.1
django-environ==0.12.0
ffmpeg==1.4
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
multidict==6.1.0
openai==1.60.1
propcache==0.2.1
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydantic==2.10.6
pydantic_core==2.27.2
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.0
tomli==2.0.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.2
yarl==1.18.3
yt-dlp==2025.1.26
